pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["mr-mapache <eric.m.cardozo@mi.unc.edu.ar>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
redis = "^5.2.1"
pydantic-settings = "^2.7.1"
fast-depends = "^2.4.12"
asyncpg = "^0.30.0"


[tool.poetry.group.tests.dependencies]
pytest-asyncio = "^0.25.1"
httpx = "^0.28.1"


[tool.poetry.group.server.dependencies]
uvicorn = "^0.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
